+++ android/app/src/main/java/app/plusone/guestlist/MainActivity.java
@@ -1,5 +1,14 @@
 package app.plusone.guestlist;
 
+import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {}
+public class MainActivity extends BridgeActivity {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        // Local plugins must be registered before the bridge starts (N5: see PushConfigPlugin).
+        registerPlugin(PushConfigPlugin.class);
+        super.onCreate(savedInstanceState);
+    }
+}